♻️ WeatherAppを別ファイルに
